--- eng/server.py
@@ -1,2 +1,0 @@
-def serve(options):
-  pass
